requirements.txt
pandas>=1.5.0
google-generativeai>=0.3.0
openpyxl>=3.0.0
streamlit>=1.28.0
python-dotenv>=1.0.0

